Level Format Refactoring (Part 1)
+ Created a more efficient file format for storing the stage data (only
present for stage 1 so far)
+ Removed rendering from the Tile class and the call to it in the Level
class, in favor of rendering the tiles directly from the level. (Tile
will be restructured significantly in the final build.)

Collision still needs to be altered to take the changes into account, as
of now, both the new and old level files are loaded by the game due to
spaghetti code that still needs to be refactored. For now, they'll need
to be left in.